Move initial history to protocol (#3422)
To fix an edge case of forking then resuming

#3419
