+++ codex-rs/core/src/rollout/mod.rs
@@ -7,9 +7,9 @@ pub mod list;
 pub(crate) mod policy;
 pub mod recorder;
 
+pub use codex_protocol::protocol::SessionMeta;
 pub use recorder::RolloutRecorder;
 pub use recorder::RolloutRecorderParams;
-pub use recorder::SessionMeta;
 
 #[cfg(test)]
 pub mod tests;
